backend/eventmanager/urls.py
"""
URL configuration for Event Manager project.
"""

from django.contrib import admin
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from eventmanager.events.views import EventViewSet

# Create a router and register our viewsets with it.
router = DefaultRouter()
router.register(r'events', EventViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
    path('api-auth/', include('rest_framework.urls')),
]
